+++ src/RadixDlt.NetworkGateway.PrometheusIntegration/GatewayApiMetricObserver.cs
@@ -84,7 +84,8 @@ internal class GatewayApiMetricObserver :
     IExceptionObserver,
     ICoreNodeHealthCheckerObserver,
     ISubmissionServiceObserver,
-    IPreviewServiceObserver,
+    ITransactionPreviewServiceObserver,
+    ITransactionOutcomeServiceObserver,
     ILedgerStateQuerierObserver,
     ISubmissionTrackingServiceObserver,
     ISqlQueryObserver
@@ -153,6 +154,27 @@ internal class GatewayApiMetricObserver :
         .CreateCounter(
             "ng_construction_transaction_preview_error_count",
             "Number of transaction preview errors",
+            new CounterConfiguration { LabelNames = new[] { "target_node" } }
+        );
+
+    private static readonly Counter _transactionOutcomeRequestCount = Metrics
+        .CreateCounter(
+            "ng_transaction_outcome_request_count",
+            "Number of transaction outcome requests",
+            new CounterConfiguration { LabelNames = new[] { "target_node" } }
+        );
+
+    private static readonly Counter _transactionOutcomeSuccessCount = Metrics
+        .CreateCounter(
+            "ng_transaction_outcome_success_count",
+            "Number of transaction outcome successes",
+            new CounterConfiguration { LabelNames = new[] { "target_node" } }
+        );
+
+    private static readonly Counter _transactionOutcomeErrorCount = Metrics
+        .CreateCounter(
+            "ng_transaction_outome_error_count",
+            "Number of transaction outcome errors",
             new CounterConfiguration { LabelNames = new[] { "target_node" } }
         );
 
@@ -257,23 +279,44 @@ internal class GatewayApiMetricObserver :
         return ValueTask.CompletedTask;
     }
 
-    ValueTask IPreviewServiceObserver.PreHandlePreviewRequest(GatewayModel.TransactionPreviewRequest request, string targetNode)
+    ValueTask ITransactionPreviewServiceObserver.PreHandlePreviewRequest(GatewayModel.TransactionPreviewRequest request, string targetNode)
     {
         _transactionPreviewRequestCount.WithLabels(targetNode).Inc();
 
         return ValueTask.CompletedTask;
     }
 
-    ValueTask IPreviewServiceObserver.PostHandlePreviewRequest(GatewayModel.TransactionPreviewRequest request, string targetNode, GatewayModel.TransactionPreviewResponse response)
+    ValueTask ITransactionPreviewServiceObserver.PostHandlePreviewRequest(GatewayModel.TransactionPreviewRequest request, string targetNode, GatewayModel.TransactionPreviewResponse response)
     {
         _transactionPreviewSuccessCount.WithLabels(targetNode).Inc();
 
         return ValueTask.CompletedTask;
     }
 
-    ValueTask IPreviewServiceObserver.HandlePreviewRequestFailed(GatewayModel.TransactionPreviewRequest request, string targetNode, Exception exception)
+    ValueTask ITransactionPreviewServiceObserver.HandlePreviewRequestFailed(GatewayModel.TransactionPreviewRequest request, string targetNode, Exception exception)
     {
         _transactionPreviewErrorCount.WithLabels(targetNode).Inc();
+
+        return ValueTask.CompletedTask;
+    }
+
+    ValueTask ITransactionOutcomeServiceObserver.PreHandleOutcomeRequest(long stateVersion, string targetNode)
+    {
+        _transactionOutcomeRequestCount.WithLabels(targetNode).Inc();
+
+        return ValueTask.CompletedTask;
+    }
+
+    ValueTask ITransactionOutcomeServiceObserver.PostHandleOutcomeRequest(long stateVersion, string targetNode, GatewayModel.TransactionCommittedOutcomeResponse response)
+    {
+        _transactionOutcomeSuccessCount.WithLabels(targetNode).Inc();
+
+        return ValueTask.CompletedTask;
+    }
+
+    ValueTask ITransactionOutcomeServiceObserver.HandleOutcomeRequestFailed(long stateVersion, string targetNode, Exception exception)
+    {
+        _transactionOutcomeErrorCount.WithLabels(targetNode).Inc();
 
         return ValueTask.CompletedTask;
     }